mn10300: Re-write move patterns.
Use the "D" and "A" constraints, and the enabled attribute to
unify all ofthe integer move patterns.  Delete the fake double
word move patterns; let the middle-end generate subregs as required.

Unfortunately, this somehow exposes a register pressure problem
with the udivmod pattern.  This is properly fixed with subsequent
patches that expose the MDR register.

In the meantime it is highly desirable to to preserve bisect-ability
of the patch series, so disable this pattern for AM30.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@169004 138bc75d-0d04-0410-961f-82ee72b054a4